Move list specific tests from order details view model to order list cell view model
